+++ bin/proving-service/src/commands/proxy.rs
@@ -2,7 +2,7 @@ use clap::Parser;
 use pingora::{
     apps::HttpServerOptions,
     prelude::{Opt, background_service},
-    server::Server,
+    server::{Server, configuration::ServerConf},
     services::listening::Service,
 };
 use pingora_proxy::http_proxy_service;
@@ -57,7 +57,15 @@ impl StartProxy {
             check_port_availability(metrics_port, "Metrics")?;
         }
 
-        let mut server = Server::new(Some(Opt::default())).map_err(|err| err.to_string())?;
+        let mut conf = ServerConf::new().ok_or(ProvingServiceError::PingoraConfigFailed(
+            "Failed to create server conf".to_string(),
+        ))?;
+        conf.grace_period_seconds = Some(self.proxy_config.grace_period_seconds);
+        conf.graceful_shutdown_timeout_seconds =
+            Some(self.proxy_config.graceful_shutdown_timeout_seconds);
+
+        let mut server = Server::new_with_opt_and_conf(Some(Opt::default()), conf);
+
         server.bootstrap();
 
         info!("Proxy starting with workers: {:?}", self.workers);