+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Allow NOOP transactions and state-updating transactions against the same account in the same block (#1393).
 - Add `bench-prover` crate to benchmark proving times (#1378)
 - Implement map in transaction kernel library (#1396).
+- Added shutdown configuration options to the `miden-proving-service` proxy (#1405).
 
 ## 0.9.1 (2025-05-30)